interface for mocking, one big fix, and several small fixes
lint fixes and others. Major fix for signal handling.

Introduced an interface for the bus operations for mocking.

The  bus mock works, but it's kind of hacky, especially mocking the signal sending.

Made the dbus connection a member of the Operations interface implementation in the real implementation. So, this allows *all* dbus calls go through the mock and real implementations.